Abstracted api, added jwt file
The JWT helper is courtesy of https://github.com/rmcdaniel/angular-codeigniter-seed/blob/master/api/application/helpers/jwt_helper.php
I merely copied the thing to use it in my tests etc. It's quite possible that I will abstract it later down the line, but I'll use it for testing for a while.